types: pin the stubs and stop chasing them
CI went red on the strict-mypy commit with thirteen errors that do not
reproduce locally. Running the same tree against a second (numpy,
pandas-stubs) pairing gave fourteen errors, only four of them shared with
CI's thirteen. The error set is a function of the stub and numpy versions,
not of the code, and no amount of per-line ignores converges on that.

So: pin pandas-stubs exactly in the dev extra, and turn warn_unused_ignores
off. That flag was the direct cause -- an ignore that is load-bearing under
one pairing is "unused" under the next, and this package supports
pandas>=2.0 and numpy across a major version boundary, so the necessary set
genuinely differs between supported environments. It joins warn_return_any
and disallow_any_generics as a check that asks about upstream's typing
rather than about us.

The twelve errors themselves were mostly worth fixing on their own terms,
which is the part that justified the detour:

  prepare     DatetimeIndex.view(int64) is deprecated in pandas 2.2+;
              astype("int64") is the supported spelling for the same
              reinterpretation. Three new columns move off .loc[:, col],
              which was never doing anything a plain assignment does not.
  inference   to_numpy(dtype=float) on series whose element type was being
              inferred from the datetime index.
  variants    the same.
  placebo     .values can be an ExtensionArray; the ufunc wants an ndarray,
              which is what the line already assumed.
  mlscm       synth_all was an ndarray and then a Series built from it; two
              things, now two names.
  normalise   GroupBy.quantile takes an array, not a list (both call sites).
  bands       np.issubdtype cannot read an ExtensionDtype at all. Replaced
              with is_datetime64_dtype -- not ..._any_dtype, because the
              numpy check was False for a tz-aware column and those must
              keep going through to_datetime_coerced.
  multisite   pivot_table wants a key, not the Index object.

Verified against both pairings: numpy 1.26 with pandas-stubs 3.0.3, and
numpy 2.4 with the pinned 2.3.2. Clean under each.

+++ src/normet/causal/inference.py
@@ -89,7 +89,7 @@ def conformal_effect_interval(
         raise ValueError("No post-period observations.")
 
     att = float(post.mean())
-    all_vals = eff.to_numpy()
+    all_vals = eff.to_numpy(dtype=float)
     n_total = len(all_vals)
     if n_total <= n_post:
         raise ValueError("Need pre-period observations to build a null distribution.")
@@ -157,8 +157,8 @@ def rmspe_ratio_test(
         post = eff_series[eff_series.index.to_series() >= cutoff_ts].dropna()
         if pre.empty or post.empty:
             return np.nan
-        pre_rmspe = float(np.sqrt(np.mean(pre.to_numpy() ** 2)))
-        post_rmspe = float(np.sqrt(np.mean(post.to_numpy() ** 2)))
+        pre_rmspe = float(np.sqrt(np.mean(pre.to_numpy(dtype=float) ** 2)))
+        post_rmspe = float(np.sqrt(np.mean(post.to_numpy(dtype=float) ** 2)))
         return post_rmspe / pre_rmspe if pre_rmspe > 0 else np.nan
 
     treated_df = placebo_space_out["treated"]
